improve selection of pack files for pinning in the open pack file cache
If two given pack files contain the same amount of requested commits,
prefer the pack file which contains the larger amount of total objects
to increase our chances of full history traversal.

Also, require the first element of the list of requested commits to be
present in the pack file since the caller will very likely try to process
this commit immediately when we return.

This fixes cases where offloading of commit coloring to got-read-pack was
ineffective, causing us to take the slow path instead.

ok op@