regress: git: don't load global/home configs
When running git(1) commands as part of the test suite,
disallow git(1) from loading either the global or home
gitconfig file.

This has the ability to mess with the test and how
they're setup.  Each test already sets its own config
so this is the preferred option anyway.

OK stsp@