fix an issue where histedit would not merge added files correctly
When a file is added during a merge and also already exists on disk as
an unversioned file, we mistakenly did not put the file onto the list
of changed files which histedit includes in the next commit. This could
result in commits which only add new files being dropped as no-op commits.

If future commits depend on the file being present this would result in
errors during the histedit operation. Otherwise, the file would no longer
exist in the edited history.

ok op@