gotwebd: rename LOGIN token to separate namespace
There's differences in how yacc(1) works on BSD vs non-BSD
systems.  As such, the generated code can often clash with
tokens which are defined elsewhere.

In this case, LOGIN from gotwebd clashes with LOGIN defined
in gotwebd.h

Simple solution here is to rename the LOGIN token in gotwebd
to be GOTWEBD_LOGIN

OK @stsp, @op2