fix bogus 'got histedit' error when deleting files which do not exist on disk
Fix a situation where an attempt to create a blob fails because the
file does not exist on disk. This blob is only created to determine
whether its content already exists in the repository, to avoid loss
of uncommitted file contents. If there is no file on disk then no
content needs to be preserved so we can just report the file as
deleted and proceed with the merge.

Problem reported by bentley@
Test case based on a reproduction script provided by bentley@